+++ src/components/CartItem.vue
@@ -3,8 +3,8 @@ tr
   td.align-middle {{ item.productName }}
   td.align-middle.text-right ${{ item.price }}
   td.align-middle.text-right
-    b-form-input.text-right.qty(@change='quantityChange', v-model.number='item.quantity', name='quantity', type='number', maxlength='2', min='1', max='10', step='1', aria-describedby='quantityFeedback', :class='{ "is-invalid": $v.item.quantity.$error }')
-    b-form-invalid-feedback(id='quantityFeedback', v-if='$v.item.quantity.$invalid') Value between 1 and 10
+    input.form-control.text-right.qty(@change='quantityChange', v-model.number='item.quantity', name='quantity', type='number', maxlength='2', min='1', max='10', step='1', aria-describedby='quantityFeedback', :class='{ "is-invalid": $v.item.quantity.$error }')
+    .invalid-feedback#quantityFeedback Value between 1 and 10
   td.align-middle.text-right
       strong ${{ item.total }}
   td.align-middle