+++ vue.config.js
@@ -15,8 +15,8 @@ module.exports = {
         options['transformAssetUrls'] = {
           img: 'src',
           image: 'xlink:href',
-          'b-img': 'src',
-          'b-img-lazy': ['src', 'blank-src'],
+          // 'b-img': 'src',
+          // 'b-img-lazy': ['src', 'blank-src'],
           // 'b-card': 'img-src',
           // 'b-card-img': 'img-src',
           // 'b-card-img-lazy': ['src', 'blank-src'],